最終版

+++ app/src/server.ts
@@ -21,6 +21,8 @@ async function connection() {
 
 // テーブル
 async function createTable(db: Database) {
+    // drop
+    await db.run('DROP TABLE IF EXISTS stock;');
     await db.run('CREATE TABLE IF NOT EXISTS stock (name STRING, amount INTEGER, sales REAL);');
     //const result = await db.get('select * from sqlite_master;');
     //console.log(result);
@@ -42,7 +44,9 @@ connection().then(async (db) => {
     app.listen(port, () => {
         console.log(`Example app listening at http://localhost:${port}`);
     });
-    
+    const error_message = { "message": "ERROR" };
+
+
     // 在庫の更新, 作成
     /**
      * POST /v1/stocks
@@ -59,17 +63,30 @@ connection().then(async (db) => {
      * same as request body
     **/
     app.post('/stocks', async (req: express.Request, res: express.Response) => {
-        const stock = req.body;
+        // リクエストボディの取得・バリデーション
+        const stocks = req.body;
+        res.setHeader('Location', `/v1/stock/${stocks.name}`);
+        if (!stocks.name) {
+            return res.status(400).json(error_message);
+        }
+        // int型かつ正の整数であるかのチェック
+        if (stocks.amount && (!Number.isInteger(stocks.amount) || stocks.amount < 0)) {
+            return res.status(400).json(error_message);
+        }
+        if (!stocks.amount) {
+            stocks.amount = 1;
+        }
+        // DBへの接続
         const db = await connection();
-        const result = await db.get('SELECT amount FROM stock WHERE name = ?', [stock.name]);
+        const result = await db.get('SELECT amount FROM stock WHERE name = ?', [stocks.name]);
         // もし在庫があれば更新
         if (result) {
-            await db.run('UPDATE stock SET amount = amount + ? WHERE name = ?', [stock.amount, stock.name]);
+            await db.run('UPDATE stock SET amount = amount + ? WHERE name = ?', [stocks.amount, stocks.name]);
         } else {
-            await db.run('INSERT INTO stock (name, amount) VALUES (?, ?)', [stock.name, stock.amount]);
+            await db.run('INSERT INTO stock (name, amount, sales) VALUES (?, ?, ?)', [stocks.name, stocks.amount, 0]);
         }
-        res.setHeader('Location', `/v1/stock/${stock.name}`);
-        res.status(201).json(stock);
+        // レスポンスボディの返却
+        return res.status(201).json(stocks);
     });
     // 在庫チェック
     /**
@@ -80,18 +97,31 @@ connection().then(async (db) => {
      * :nameが指定されていない場合
      * 全ての商品の在庫の数を:nameで昇順sortして
      * {[name]: [amount], [name]: [amount], ...}で返す
+     * 例: {"apple": 1, "banana": 2, "orange": 0}
      * 在庫が0のものは含まない
     **/
     app.get('/stocks/:name?', async (req: express.Request, res: express.Response) => {
+        // リクエストパラメータの取得・バリデーション
         const name = req.params.name;
+        // nameのバリデーション
+        if (name && typeof name !== 'string') {
+            return res.status(400).json(error_message);
+        }
         const db = await connection();
         let result: Stock[] = [];
         if (name) {
             result = await db.all('SELECT name, amount FROM stock WHERE name = ?', [name]);
+            return res.json(result);
         } else {
             result = await db.all('SELECT name, amount FROM stock WHERE amount > 0 ORDER BY name ASC');
+            // 形式を変換
+            // 例: {"apple": 1, "banana": 2, "orange": 0}
+            const tmp: { [key: string]: number } = {};
+            result.forEach((stock) => {
+                tmp[stock.name] = stock.amount;
+            });
+            return res.json(tmp);
         }
-        res.json(result);
     }
     );
     // 販売
@@ -111,18 +141,36 @@ connection().then(async (db) => {
      * request bodyと同じ
      **/
     app.post('/sales', async (req: express.Request, res: express.Response) => {
+        // リクエストボディの取得・バリデーション
         const sales = req.body;
+        res.setHeader('Location', `/v1/sales/${sales.name}`);
+        if (!sales.name) {
+            return res.status(400).json(error_message);
+        }
+        // int型かつ正の整数であるかのチェック
+        if (sales.amount && (!Number.isInteger(sales.amount) || sales.amount < 0)) {
+            return res.status(400).json(error_message);
+        }
+        // int型かつ正の整数であるかのチェック
+        if (sales.price && (!Number.isInteger(sales.price) || sales.price < 0)) {
+            return res.status(400).json(error_message);
+        }
+        // amountの初期化
+        let amount = 1;
+        // もしamountが指定されていればamountを更新
+        if (sales.amount) {
+            amount = sales.amount;
+        }
         const db = await connection();
         const result = await db.get('SELECT amount FROM stock WHERE name = ?', [sales.name]);
-        if (result.amount < sales.amount) {
-            res.status(400).json({ error: '在庫が足りません' });
+        if (!result || result.amount < amount) {
+            return res.status(400).json(error_message);
         } else {
-            await db.run('UPDATE stock SET amount = amount - ? WHERE name = ?', [sales.amount, sales.name]);
+            await db.run('UPDATE stock SET amount = amount - ? WHERE name = ?', [amount, sales.name]);
             if (sales.price) {
-                await db.run('UPDATE stock SET sales = sales + ? WHERE name = ?', [sales.price * sales.amount, sales.name]);
+                await db.run('UPDATE stock SET sales = sales + ? WHERE name = ?', [sales.price * amount, sales.name]);
             }
-            res.setHeader('Location', `/v1/sales/${sales.name}`);
-            res.status(201).json(sales);
+            return res.status(201).json(sales);
         }
     }
     );
@@ -133,14 +181,21 @@ connection().then(async (db) => {
      * 
      * http response body
      * {
-     * "sales": 0 (売り上げの合計額, 小数第二位まで)
+     * "sales": 0 (売り上げの合計額, 小数第二位まで, 整数の場合は小数第一位まで)
      * }
      * 
      **/
     app.get('/sales', async (req: express.Request, res: express.Response) => {
+        // リクエストパラメータの取得・バリデーション
+        // もしパラメータがあればエラー
+        if (Object.keys(req.query).length !== 0) {
+            return res.status(400).json(error_message);
+        }
         const db = await connection();
-        const result = await db.get('SELECT SUM(sales) AS sales FROM stock');
-        res.json(result);
+        const sales = await db.get('SELECT SUM(sales) AS sales FROM stock');
+        // 小数であれば小数第二位まで、整数であれば小数第一位まで
+        const result = { sales: sales.sales.toFixed(sales.sales % 1 === 0 ? 1 : 2) };
+        return res.json(result);
     }
     );
 
@@ -149,9 +204,14 @@ connection().then(async (db) => {
      * DELETE /v1/stocks
     **/
     app.delete('/stocks', async (req: express.Request, res: express.Response) => {
+        // リクエストパラメータの取得・バリデーション
+        // もしパラメータがあればエラー
+        if (Object.keys(req.query).length !== 0) {
+            return res.status(400).json(error_message);
+        }
         const db = await connection();
         await db.run('DELETE FROM stock');
-        res.status(204).end();
+        return res.status(204).end();
     }
     );
 });
